auth commit

+++ src/main/kotlin/com/learning/spring_boot_learning/controllers/NoteController.kt
@@ -5,6 +5,7 @@ import com.learning.spring_boot_learning.database.model.Note
 import com.learning.spring_boot_learning.database.repository.NoteRepository
 import io.jsonwebtoken.lang.Objects
 import org.bson.types.ObjectId
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -18,7 +19,8 @@ import java.time.Instant
 @RestController
 @RequestMapping("/notes")
 class NoteController(
-    private val repository: NoteRepository
+    private val repository: NoteRepository,
+    private val noteRepository: NoteRepository
 ) {
 
     data class NoteRequest(
@@ -40,6 +42,7 @@ class NoteController(
     fun save(
         @RequestBody body: NoteRequest
     ): NoteResponse{
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val note = repository.save(
                 Note(
                     id = body.id?.let{ ObjectId(it)}?:ObjectId.get(),
@@ -47,7 +50,7 @@ class NoteController(
                     content = body.content,
                     color = body.color,
                     createdAt = Instant.now(),
-                    ownerId = ObjectId()
+                    ownerId = ObjectId(ownerId)
                 )
             )
 
@@ -56,17 +59,25 @@ class NoteController(
 
 
     @GetMapping
-    fun findByOwnerId(
-        @RequestParam(required = true) ownerId: String): List<NoteResponse>{
+    fun findByOwnerId():List<NoteResponse>{
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+
         return repository.findByOwnerId(ObjectId(ownerId)).map {
-            it.toResponse()
-        }
+        it.toResponse()
+    }
     }
 
 
     @DeleteMapping(path = ["/{id}"])
     fun deleteById(@PathVariable id: String){
-        repository.deleteById(ObjectId(id))
+        val note = noteRepository.findById(ObjectId(id)).orElseThrow{
+            IllegalArgumentException("Note not fund")
+        }
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        if (note.ownerId.toHexString() == ownerId){
+            repository.deleteById(ObjectId(id))
+        }
+
     }
 
 }
